requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.6.3
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.4
decorator==4.3.0
dj-database-url==0.4.2
Django==1.11.5
django-contrib-comments==1.9.0
django-debug-toolbar==1.5
django-extensions==2.1.3
django-js-asset==1.1.0
django-static-precompiler==1.6
django-tagging==0.4.6
django-webpack-loader==0.5.0
django-xmlrpc==0.1.8
djangorestframework==3.6.3
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.25.0
google-cloud-storage==1.2.0
google-resumable-media==0.1.1
googleapis-common-protos==1.5.2
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
mots-vides==2015.5.11
oauth2client==4.1.2
olefile==0.44
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==4.1.1
prompt-toolkit==2.0.7
protobuf==3.3.0
psycopg2==2.7.3.1
ptyprocess==0.6.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
Pygments==2.2.0
pyparsing==2.3.0
pytz==2017.2
PyYAML==3.12
regex==2018.11.7
requests==2.18.1
rsa==3.4.2
six==1.10.0
sqlparse==0.2.3
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
whitenoise==3.0
yamjam==0.1.7
